clean before merge
